+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 0.10.1
 * + tests for get_records always returns list of tuples
-* resolve many bugs created by select() changes in v0.10.0
+* resolve many bugs created by get_records() changes in v0.10.0
 
 # 0.10.0
 * BREAKING: select() now *always* returns list of tuples eg: [(0,1)]